refactor(ui): migrate kernel constants to types library
