+++ apps/ui/app/components/chat/kernel-selector.tsx
@@ -1,9 +1,9 @@
 import type { KernelProvider } from '@taucad/types';
+import { kernelConfigurations } from '@taucad/types/constants';
 import { Button } from '#components/ui/button.js';
 import { HoverCard, HoverCardContent, HoverCardTrigger } from '#components/ui/hover-card.js';
 import { Badge } from '#components/ui/badge.js';
 import { SvgIcon } from '#components/icons/svg-icon.js';
-import { kernelOptions } from '#constants/kernel.constants.js';
 import { cn } from '#utils/ui.utils.js';
 
 export type KernelSelectorProperties = {
@@ -19,7 +19,7 @@ export function KernelSelector({
 }: KernelSelectorProperties): React.JSX.Element {
   return (
     <div className="grid grid-cols-3 gap-3 lg:grid-cols-3">
-      {kernelOptions.map((option) => (
+      {kernelConfigurations.map((option) => (
         <HoverCard key={option.id}>
           <HoverCardTrigger asChild>
             <Button
