+++ apps/ui/app/routes/builds_.$id/chat-controls.tsx
@@ -2,6 +2,7 @@ import { Clipboard, Download, GalleryThumbnails, ImageDown, Menu } from 'lucide-
 import { useCallback, useEffect, useMemo } from 'react';
 import { useSelector, useActorRef } from '@xstate/react';
 import type { ExportFormat } from '@taucad/types';
+import { fileExtensionFromExportFormat } from '@taucad/types/constants';
 import { BoxDown } from '#components/icons/box-down.js';
 import { Button } from '#components/ui/button.js';
 import { useBuildSelector } from '#hooks/use-build.js';
@@ -13,7 +14,6 @@ import { ComboBoxResponsive } from '#components/ui/combobox-responsive.js';
 import { downloadBlob } from '#utils/file.utils.js';
 import { screenshotRequestMachine } from '#machines/screenshot-request.machine.js';
 import { exportGeometryMachine } from '#machines/export-geometry.machine.js';
-import { extensionFromFormat } from '#constants/kernel.constants.js';
 
 type ViewerControlItem = {
   id: string;
@@ -74,7 +74,7 @@ export function ChatControls(): React.JSX.Element {
 
   const handleExport = useCallback(
     async (filename: string, format: ExportFormat) => {
-      const fileExtension = extensionFromFormat[format];
+      const fileExtension = fileExtensionFromExportFormat[format];
       const filenameWithExtension = `${filename}.${fileExtension}`;
       toast.promise(
         new Promise<Blob>((resolve, reject) => {
