+++ apps/ui/app/routes/builds_.$id/command-palette.tsx
@@ -15,6 +15,7 @@ import { useSelector, useActorRef } from '@xstate/react';
 import { Link, useNavigate } from 'react-router';
 import { useAuthenticate } from '@daveyplate/better-auth-ui';
 import type { ExportFormat } from '@taucad/types';
+import { fileExtensionFromExportFormat } from '@taucad/types/constants';
 import { Button } from '#components/ui/button.js';
 import { useBuildSelector } from '#hooks/use-build.js';
 import { toast } from '#components/ui/sonner.js';
@@ -34,7 +35,6 @@ import {
 import { useKeydown } from '#hooks/use-keydown.js';
 import { KeyShortcut } from '#components/ui/key-shortcut.js';
 import type { KeyCombination } from '#utils/keys.utils.js';
-import { extensionFromFormat } from '#constants/kernel.constants.js';
 import { SvgIcon } from '#components/icons/svg-icon.js';
 import { Format3D } from '#components/icons/format-3d.js';
 
@@ -108,7 +108,7 @@ export function CommandPalette({ isOpen, onOpenChange }: CommandPalettePropertie
 
   const handleExport = useCallback(
     async (filename: string, format: ExportFormat) => {
-      const fileExtension = extensionFromFormat[format];
+      const fileExtension = fileExtensionFromExportFormat[format];
       const filenameWithExtension = `${filename}.${fileExtension}`;
       toast.promise(
         new Promise<Blob>((resolve, reject) => {
