+++ libs/types/src/constants/index.ts
@@ -1,9 +1,10 @@
 /* eslint-disable no-barrel-files/no-barrel-files -- package entry point. */
 export * from '#constants/cad.constants.js';
 export * from '#constants/code.constants.js';
-export * from '#constants/manufacturing.constants.js';
-export * from '#constants/message.constants.js';
-export * from '#constants/tool.constants.js';
+export * from '#constants/file.constants.js';
 export * from '#constants/id.constants.js';
 export * from '#constants/kernel.constants.js';
+export * from '#constants/manufacturing.constants.js';
+export * from '#constants/message.constants.js';
 export * from '#constants/model.constants.js';
+export * from '#constants/tool.constants.js';
