+++ libs/types/src/types/file.types.ts
@@ -1,0 +1,3 @@
+import type { exportFormats } from '#constants/file.constants.js';
+
+export type ExportFormat = (typeof exportFormats)[number];
