+++ libs/types/src/types/index.ts
@@ -4,6 +4,7 @@ export type * from '#types/cad.types.js';
 export type * from '#types/chat.types.js';
 export type * from '#types/code.types.js';
 export type * from '#types/constant.types.js';
+export type * from '#types/file.types.js';
 export type * from '#types/graphics.types.js';
 export type * from '#types/id.types.js';
 export type * from '#types/kernel.types.js';
